fix: Change SpO2 parameter type from Int to Float
Fixed type mismatch error in SpO2Card composable:
- Changed spO2 parameter from Int to Float
- Updated default value from 0 to 0f

Fixes: 'Argument type mismatch' build error

+++ src/main/kotlin/com/fitness/app/presentation/ring/screens/RingSetupScreen.kt
@@ -624,7 +624,7 @@ private fun ConnectedContent(
             // Blood Oxygen (SpO2)
             item {
                 SpO2Card(
-                    spO2 = uiState.ringData?.spO2 ?: 0,
+                    spO2 = uiState.ringData?.spO2 ?: 0f,
                     isMeasuring = uiState.ringData?.spO2Measuring == true
                 )
             }
@@ -771,7 +771,7 @@ private fun BloodPressureCard(systolic: Int, diastolic: Int, isMeasuring: Boolea
 }
 
 @Composable
-private fun SpO2Card(spO2: Int, isMeasuring: Boolean) {
+private fun SpO2Card(spO2: Float, isMeasuring: Boolean) {
     Surface(
         modifier = Modifier.fillMaxWidth(),
         shape = RoundedCornerShape(16.dp),
